fix: remove panic when assembling empty model set
Before, if you attempted to assemble an empty set of models (i.e. a
directory with no matching smithy/json files), we panicked.

This is handled in TryFrom<ModelAssembler> by explicitly
checking whether the `Vec<Model>` is empty.

Now, in that case you get a warning logged and an default model returned
as Ok().